Code cleanup and ChestGui improvements

+++ src/com/massivecraft/massivecore/ps/PS.java
@@ -506,7 +506,7 @@ public final class PS implements Serializable, Comparable<PS>
 	public static Integer calcBlockCoord(Double location, Integer block, Integer chunk)
 	{
 		if (block != null) return block;
-		if (location != null) return (int) Location.locToBlock(location);
+		if (location != null) return Location.locToBlock(location);
 		if (chunk != null) return chunk * 16;
 		return null;
 	}
